pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "saltpaper"
version = "0.0.6a0"
authors = [
    { name="Lauren Kinder", email="mail@hermod.uk" },
]
description = "A small game-engine/framework built on top of pygame"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "pygame-ce>=2.4.0",
    "numpy>=1.24.0",
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
license = "MIT"
license-files = ["LICEN[CS]E*"]

[tool.setuptools.packages.find]
where = ["."]
include = ["saltpaper*"]

[tool.setuptools]
include-package-data = true
